context api partialy applied - deleted method improved

+++ src/Components/TaskList.js
@@ -1,14 +1,17 @@
 
-import React from "react";
+import React, {useContext} from "react";
 import { Container, Stack, Row, Col, Button } from "react-bootstrap";
 import firebaseApp from "../Config/firebase";
 import { getFirestore, doc, updateDoc } from "firebase/firestore";
 import { getStorage, ref, deleteObject } from "firebase/storage";
+import TaskListContext from "../Contexts/TaskListContext";
 const firestore = getFirestore(firebaseApp);
 const storage = getStorage(firebaseApp);
 
 
-const TaskList = ({tasks, setTasks, globalUserEmail})=>{
+const TaskList = ({tasks, setTasks})=>{
+
+    const context = useContext(TaskListContext);
 
     async function deleteTask (tIdtoDelete){
         
@@ -17,12 +20,17 @@ const TaskList = ({tasks, setTasks, globalUserEmail})=>{
         //borrar el archivo del storage
         const taskToDelete = tasks.filter((t)=> t.itemId == tIdtoDelete);
         console.log("taskToDelete: ", taskToDelete);
-        const urlToDelete = taskToDelete[0].url.downloadUrl;        
+        const urlToDelete = taskToDelete[0].downloadUrl;        
         console.log("URL: ", urlToDelete);
-        const fileRef = ref(storage, urlToDelete);
-        await deleteObject(fileRef);
+
+        const firebaseUrl = "https://firebasestorage";
+
+        if(urlToDelete.includes(firebaseUrl)){
+            const fileRef = ref(storage, urlToDelete);
+            await deleteObject(fileRef);
+        } 
         //actualizar la base de datos
-        const docRef = doc(firestore, `usersDocs/${globalUserEmail}`);
+        const docRef = doc(firestore, `usersDocs/${context.globalUser.email}`);
         updateDoc(docRef, {tasks: [...newTasks]});  
         //actualizar el estado correspondiente
         setTasks(newTasks);
@@ -31,7 +39,9 @@ const TaskList = ({tasks, setTasks, globalUserEmail})=>{
 
     }
 
-    console.log("MAIL EN TASKLIST: ", globalUserEmail);
+    console.log("MAIL EN TASKLIST: ", context.globalUser.email);
+
+
     return (
         <Container>
             <Stack>
@@ -43,7 +53,7 @@ const TaskList = ({tasks, setTasks, globalUserEmail})=>{
                             <Row key={t.id}>
                                 <Col  >{t.description}</Col>
                                 <Col>
-                                <a href={t.url.downloadUrl}><Button>Ver</Button></a></Col>
+                                <a href={t.downloadUrl}><Button>Ver</Button></a></Col>
                                 <Col><Button variant="danger" onClick={()=>deleteTask(t.itemId)}>Eliminar</Button></Col>
                             </Row>
                             <hr/>
